requirements.txt
flask
flask-cors
matplotlib
numpy
pyyaml
pillow
